Refactor app to use Redis instead of local files.
Heroku has a VERY ephemeral filesystem, so using local files for the modified
story text and and list of changes turned out to be a terrible way to work
since it meant changes would never be recorded! Potential solutions were
"switch away from Heroku" (AWS or Digital Ocean?) or "use one of Heroku's
solutions for persistent data storage" (Postgres or Redis). Decided to stick
with Heroku bc a) I know and use it, and b) it has a free tier and the other
two do not. Went with Redis bc it's more lightweight and I wanted to learn how
it works.