Adding initial JS
This is a change back from when I initially worked on this in 2016.
It's basically meaningless to me (although I understand why I was
doing it) and I have no idea how much more work is needed. As-is, it
simply prints "None" in place of the story body when running and I
don't know what more is needed. Deal with this later.