Save original text to redis during initialization
This would fail the very first time if you were starting with a new
and empty redis db. By default, save the original text to the
mutated key as well as the original key on startup so fetching the
tokens doesn't need to worry about which key the tokens should be
coming from.