Everything working on 2.7 again
Next up: try to upgrade to 3.8. I'm currently using pyenv and pyenv-
virtualenv to switch between two sandboxed virtualenvs with different
Python versions and have pip installed everything necessary in both.
This existing version of requirements.txt is therefore 2.7-specific,
but if all goes well the next update to that file should be for all
necessary 3.8 dependencies instead.

I also added .python-version to gitignore. This is a file that pyenv
uses to keep track of the local version of Python (again, currently
2.7.17). However, the file is less useful to track now that I've
figured out pyenv-virtualenv, which overwrites the Python version in
the file with the name of your virtualenv. Not useful to anyone else.