Add runtime.txt for Heroku
I'm assuming this is now explicitly needed for my app since my app is
a) so old (previously 2.7.11 according to my failed attempt to deploy
to Heroku?) and b) simultaneously migrating to a new type of stack on
Heroku. Normally for new python apps Heroku will implicitly default
to python-3.6.9, although you can add a runtime.txt to choose an
alternative supported version, so I'm trying that here to match my
recent upgrades to a supported version of Python.