Add nltk.txt
There are a bunch of poorly documented ways to allow the downloading
of NLTK data (since NLTK doesn't download everything when the
library is installed) on Heroku. This is messy because typically NLTK
expects you to do the data downloads manually via their interactive
downloader (see https://www.nltk.org/data.html), which of course
doesn't really work on Heroku.

Some things I've found (and some I've tried), all unsuccessfully so
far:

- Using the `post_compile` and `pre_compile` hooks that exist in
Heroku's python buildpack but are completely undocumented.
https://github.com/heroku/heroku-buildpack-python/blob/master/bin/steps/hooks/post_compile
This is what previously worked until I started doing all this work to
upgrade everything on this Heroku app and it no longer works after
moving from a cedar-14 stack to a heroku-18 stack.

- Adding a `release` line to the Procfile that manually runs the
binscript `post_compile` points to. This didn't work locally and I
didn't commit the change, but might work? Docs here:
https://devcenter.heroku.com/articles/procfile#the-release-process-type

- Using a new `nltk.txt` per an existing, but unlisted, Heroku doc:
https://devcenter.heroku.com/articles/python-nltk This did not work
locally but this commit implements this solution because it might
work on Heroku but not locally (see below).

- Manually downloading and unzipping all relevant NLTK data locally
and committing it to the repo, per
https://stackoverflow.com/a/14869451 I'm assuming this would work but
I'm currently treating it as a last resort and have not yet tried it.

A large part of the problem, as far as I can tell, is that while you
can _run_ your Heroku app locally via `heroku local` this does not
_build_ the app with the relevant buildpack. As a result, I cannot
verify the first three solutions above locally and must make a commit
to test one. I've chosen to try adding the nltk.txt since it's the
simplest, but I don't have high hopes since `post_compile` previously
worked and is also still part of python buildpack but afaict no
longer actually works.